b/pyproject.toml
[project]
name = "paperclip-adjacent"
version = "0.1.0"
dependencies = ["flask", "tinydb", "flask-cors", "pyyaml"]

[project.optional-dependencies]

[build-system]
requires = ["setuptools"]
build-backend = "setuptools.build_meta"

[tool.setuptools]
packages = ["App"]

# --- CLI COMMANDS ---
[project.scripts]
paperclip = "App.__main__:main"